requirements.txt
python-telegram-bot==22.7
SQLAlchemy==2.0.49
APScheduler==3.11.2
dateparser==1.4.0
python-dotenv==1.1.0
groq==1.1.2
pydantic==2.13.0
